Fix Project 2. Delete old images. Use new entropy function. Calculate everything in doubles (not uint8).

+++ Project 2/project_2/newcode/jzlk_ifwt2Direct.m
@@ -23,7 +23,7 @@ for ii=scale:-1:1
     xhat(1:midx1, 1:midx2) = xhatPart;
 end
 
-imagehat = uint8(xhat);
+imagehat = xhat;
 
 end
